Reorganize filesystem and artifact manager responsibilities
